% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ea.R
\name{ea}
\alias{ea}
\title{Actual evapotranspiration}
\usage{
ea(soilType, eatemp, cea, M, D, G, middelsca)
}
\arguments{
\item{soilType}{type of soil (soil or bog)}

\item{eatemp}{evapotransiration}

\item{cea}{degree-day factor, positive (resp neg) for positive (resp neg) temperature}

\item{M}{groundwater storage capacity}

\item{D}{the potential volumee of water that is needed for complete saturation at time t}

\item{G}{Volume of received moisture input of rain and snowmelt a time t}

\item{middelsca}{average snow coverage over the level zones}
}
\description{
The function \code{ea()} processes the actual evapotranspiration.
It follows the expression of Skaugen,Peerebom and Nilsson(2015):
Ea = Ep*(S+Z)/M, with Ep = cea*T, the potential evapotranspiration
}
\examples{
\dontrun{
ea()
}
}
\keyword{soilWater}

