% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excessWater.R
\name{excessWater}
\alias{excessWater}
\title{Excess of water}
\usage{
excessWater(G, Z, D, R)
}
\arguments{
\item{G}{Volume of received moisture input of rain and snowmelt a time t}

\item{Z}{Actual water volume present in the soil moisture zone}

\item{D}{the potential volumee of water that is needed for complete saturation at time t}

\item{R}{Ratio defining field capacity (fracion of D)}
}
\description{
The function \code{excessWater()} processes the excess of water.
It follows the expression of Skaugen and Onof (2013) and Skaugen,Peerebom and Nilsson(2015)
if (G(t)+Z(t))/D(t) >R : X(t) = G(t) + Z(t) -R*D(t)
if (G(t)+Z(t))/D(t) <R : X(t) = 0
}
\examples{
\dontrun{
excessWater()
}
}
\keyword{soilWater}

