% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateX.R
\name{stateX}
\alias{stateX}
\title{Processinf the soil water content}
\usage{
stateX(eatemp, cea, M, D, G, middelsca, R, Z, Gbog, Zbog)
}
\arguments{
\item{eatemp}{evapotranspiration}

\item{cea}{Degree day factor for evpotranspiration(mm/degree/day)}

\item{M}{Groundwater Storage Capacity (GSC)}

\item{D}{the potential volumee of water that is needed for complete saturation at time t}

\item{G}{Volume of received moisture input of rain and snowmelt a time t}

\item{middelsca}{average snow coverage over the level zones}

\item{R}{Ratio defining field capacity (fracion of D)}

\item{Z}{Actual water volume present in the soil moisture zone}

\item{Gbog}{Volume of received moisture input of rain and snowmelt a time t on Bogs}

\item{Zbog}{Actual water volume present on bogs}
}
\description{
The function \code{stateX()} processes the soil water content both for soils and bogs.
The process follows the expression dZ/dt= G(t) - X(t)- Ea(t) in Skaugen, Peerebom and Nilsson (2015).
There are hence three steps:
 i- it processes the actual evapotranspiration Ea(t)
 ii- it processes the soil moisture G(t)
 iii- it processes the excess water X(t)
}
\examples{
\dontrun{
stateX()
}
}
\keyword{soilWater}

